% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madhttr-package.R
\name{download_ssl_cert}
\alias{download_ssl_cert}
\title{Doewnload X.509 certificates}
\arguments{
\item{host, port, ipv4_only}{see \code{\link[openssl:download_ssl_cert]{openssl::download_ssl_cert()}}}
}
\description{
Doewnload X.509 certificates
}
