% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madhttr-package.R
\docType{package}
\name{madhttr}
\alias{madhttr}
\alias{madhttr-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/madhttr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/madhttr/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/madhttr}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/madhttr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
