% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madhttr-package.R
\name{read_har}
\alias{read_har}
\title{Read HAR objects}
\usage{
read_har(har, ...)
}
\arguments{
\item{har}{a string, list/URL or a file containing JSON HAR data}

\item{...}{additional arguments}
}
\description{
Snake-case re-exported alias for \code{\link[HARtools:readHAR]{HARtools::readHAR()}}.
}
