% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-cert.R
\name{tidy_cert}
\alias{tidy_cert}
\title{Turn an {openssl} downloaded SSL certificate into a tidy data frame}
\usage{
tidy_cert(.x)
}
\arguments{
\item{.x}{a certificate  list retrieved with \code{\link[openssl:download_ssl_cert]{openssl::download_ssl_cert()}} or
a single \code{cert} object.}
}
\description{
Turn an {openssl} downloaded SSL certificate into a tidy data frame
}
\examples{
tidy_cert(openssl::download_ssl_cert("rud.is"))
tidy_cert(openssl::download_ssl_cert("r-project.org")[[2]])
}
