% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-har.R
\name{tidy_har}
\alias{tidy_har}
\title{Tidy HAR entries}
\usage{
tidy_har(.x)
}
\arguments{
\item{a}{\code{harentry} or \code{harentries}/\code{harlog}/\code{har} object}
}
\description{
Given a single \code{harentry} or a \code{harentries} list/\code{harlog}/\code{har}, return a data frame
with selected columns.
}
\examples{
tidy_har(readRDS(system.file("extdat", "example-har.rds", package = "madhttr")))
}
