% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-res.R
\name{tidy_response}
\alias{tidy_response}
\title{Turn an {httr} response object into a tidy data frame}
\usage{
tidy_response(.x)
}
\arguments{
\item{.x}{a \code{response} object retrieved with one of the \code{\link[httr:VERB]{httr::VERB()}} functions.}
}
\description{
Turn an {httr} response object into a tidy data frame
}
\examples{
tidy_response(httr::GET("https://rud.is"))
}
