% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madhttr-package.R
\name{write_har}
\alias{write_har}
\title{Write HAR objects}
\usage{
write_har(har, file, force = TRUE, ...)
}
\arguments{
\item{har}{a string, list/URL or a file containing JSON HAR data}

\item{file, force, ...}{see \code{\link[HARtools:writeHAR]{HARtools::writeHAR()}}}
}
\description{
Snake-case re-exported alias for \code{\link[HARtools:writeHAR]{HARtools::writeHAR()}}.
}
