% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madhttr-package.R
\docType{package}
\name{madhttr}
\alias{madhttr}
\alias{madhttr-package}
\title{Tidy Helper Methods for Many Types of Unkempt Internet Metadata and Content}
\description{
The 'httr', 'openssl', and 'HARtools' packages provide methods to retrieve
rich metadata and content from internet hosts but their return objects are quite
unkempt. Methods are provided to turn these objects into tidy data frames along with
other useful helper methods which augment functionality in these packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/madhttr}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/madhttr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
